Refactor dump logic into resource/resolve
